+++ lib/rubocop/cop/rspec_cops.rb
@@ -13,6 +13,7 @@ begin
 rescue LoadError
   # Rails/HttpStatus cannot be loaded if rack/utils is unavailable.
 end
+require_relative 'rspec/rails/persistence_called_outside_example'
 
 require_relative 'rspec/align_left_let_brace'
 require_relative 'rspec/align_right_let_brace'